Gstreamer: refactor how focus point setting works
Previously, when the focus point mode is set to custom, it won't send
out region-on-interest update to Gstreamer. To fix this, I decided to
refactor things a bit to make the flow of data clear.

First, m_focusRect is removed. This value depends on focus point mode
and viewfinder resolution, and is not valid in face detection mode,
which can be confusing. To make sure that the focus rect size will still
has the same aspect ratio as the viewfinder, the size part of this rect
is replaced by m_focusRectSize.

The QRectF variant of updateRegionOfInterest() is removed and the
QVector<QRect> variant is renamed to sendRegionOfInterestEvent() to
more accurately reflects what it does. The new argument-less
updateRegionOfInterest() choose how to constructs the list of RoI based
on focus point mode. focusZones() is also updated to choose how to
construct the list of zones based on focus point mode.

Then, whenever the focus point mode is changed, the updateRegionOf
Interest() is called to make sure that the RoI in Gstreamer matches our
settings.

Finally, changing focus point mode no longer reset custom focus point
users has set. I think it doesn't make sense and it will create
confusion as this isn't documented anywhere.

[ChangeLog][CameraBin] fixed setting a custom focus point before setting
focus point mode to custom.

Change-Id: I11c6f8133d6f42d9860c88d14eec715d8fb853bc
Forwarded: no
Last-Update: 2019-10-28